+++ zest/releaser/baserelease.py
@@ -77,15 +77,17 @@ class Basereleaser:
                 "zest.releaser.tests", "pypirc_old.txt"
             )
             self.pypiconfig = pypi.PypiConfig(pypirc_old)
+            self.zest_releaser_config = pypi.ZestReleaserConfig(pypirc_config_filename=pypirc_old)
         else:
             self.pypiconfig = pypi.PypiConfig()
-        if self.pypiconfig.no_input():
+            self.zest_releaser_config = pypi.ZestReleaserConfig()
+        if self.zest_releaser_config.no_input():
             utils.AUTO_RESPONSE = True
 
     @property
     def history_format(self):
         default = "rst"
-        config_value = self.pypiconfig.history_format()
+        config_value = self.zest_releaser_config.history_format()
         history_file = self.data.get("history_file") or ""
         return utils.history_format(config_value, history_file)
 
@@ -124,8 +126,8 @@ class Basereleaser:
         self.data["headings"] = []
         self.data["history_last_release"] = ""
         self.data["history_insert_line_here"] = 0
-        default_location = self.pypiconfig.history_file()
-        fallback_encoding = self.pypiconfig.encoding()
+        default_location = self.zest_releaser_config.history_file()
+        fallback_encoding = self.zest_releaser_config.encoding()
         history_file = self.vcs.history_file(location=default_location)
         self.data["history_file"] = history_file
         if not history_file:
@@ -277,10 +279,9 @@ class Basereleaser:
                 "the existing file. This might give problems.",
                 orig_encoding,
             )
-            config = self.setup_cfg.config
             fallback_encodings = []
-            if config.has_option("zest.releaser", "encoding"):
-                encoding = config.get("zest.releaser", "encoding")
+            if "encoding" in self.zest_releaser_config:
+                encoding = self.zest_releaser_config["encoding"]
                 if encoding != orig_encoding:
                     fallback_encodings.append(encoding)
             encoding = "utf-8"
@@ -416,7 +417,7 @@ class Basereleaser:
         push_cmds = self.vcs.push_commands()
         if not push_cmds:
             return
-        default_anwer = self.pypiconfig.push_changes()
+        default_anwer = self.zest_releaser_config.push_changes()
         if utils.ask("OK to push commits to the server?", default=default_anwer):
             for push_cmd in push_cmds:
                 output = execute_command(push_cmd)
@@ -424,7 +425,7 @@ class Basereleaser:
 
     def _run_hooks(self, when):
         which_releaser = self.__class__.__name__.lower()
-        utils.run_hooks(self.setup_cfg, which_releaser, when, self.data)
+        utils.run_hooks(self.zest_releaser_config, which_releaser, when, self.data)
 
     def run(self):
         self._run_hooks("before")
@@ -440,8 +441,8 @@ class Basereleaser:
         raise NotImplementedError()
 
     def update_commit_message(self, msg):
-        prefix_message = self.pypiconfig.prefix_message()
-        extra_message = self.pypiconfig.extra_message()
+        prefix_message = self.zest_releaser_config.prefix_message()
+        extra_message = self.zest_releaser_config.extra_message()
         if prefix_message:
             msg = "%s %s" % (prefix_message, msg)
         if extra_message: