+++ zest/releaser/prerelease.py
@@ -45,7 +45,7 @@ class Prereleaser(baserelease.Basereleaser):
     def __init__(self, vcs=None):
         baserelease.Basereleaser.__init__(self, vcs=vcs)
         # Prepare some defaults for potential overriding.
-        date_format = self.pypiconfig.date_format()
+        date_format = self.zest_releaser_config.date_format()
         self.data.update(
             dict(
                 commit_msg=PRERELEASE_COMMIT_MSG,